requirement-backup.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
slowapi==0.1.9
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
wrapt==2.0.1
fastapi
uvicorn 
sqlalchemy
alembic
psycopg2-binary
slowapi
python-dotenv
pyjwt
passlib
bcrypt==4.0.1
python-multipart
email-validator
